requirements.txt
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1687884187023/work
cachetools==5.3.3
captum==0.2.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
cilog==1.3.0
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dive-into-graphs==1.1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fonttools==4.52.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
munch==4.0.0
mypy-extensions==1.0.0
networkx==3.1
numba==0.58.1
numpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_bdk7p7tnsw/croot/numpy_and_numpy_base_1682520579379/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1701735466804/work
omegaconf==2.3.0
openai==1.30.3
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1621288026676/work
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pyscf==2.5.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rdkit-pypi==2022.9.5
referencing==0.35.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1716354486713/work
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.3.2
scipy==1.10.1
shap==0.44.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.35.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
torch==1.11.0
torch-geometric==2.0.4
torch-scatter==2.1.1
torch-sparse==0.6.15
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typed-argument-parser==1.7.2
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata==2024.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
watchdog==4.0.1
wcwidth==0.2.13
zipp==3.19.0
